@@ -1,38 +1,30 @@
 package com.example.vmac.WatBot;
 
-/**
- * Created by VMac on 17/11/16.
- */
-
-import android.app.Activity;
-import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageView;
 import android.widget.TextView;
-
+import androidx.annotation.NonNull;
+import androidx.recyclerview.widget.RecyclerView;
 import com.bumptech.glide.Glide;
-
+import com.bumptech.glide.load.engine.DiskCacheStrategy;
+import com.bumptech.glide.request.RequestOptions;
 import java.util.ArrayList;
 
-
-public class ChatAdapter extends RecyclerView.Adapter<RecyclerView.ViewHolder> {
-
-    protected Activity activity;
-    private int SELF = 100;
-    private ArrayList<Message> messageArrayList;
-
+public class ChatAdapter extends RecyclerView.Adapter<ChatAdapter.ViewHolder> {
+    private static final int SELF = 100;
+    private static final int WATSON = 200;
+    private final ArrayList<Message> messageArrayList;
 
     public ChatAdapter(ArrayList<Message> messageArrayList) {
-        this.messageArrayList = messageArrayList;
-
+        this.messageArrayList = messageArrayList != null ? messageArrayList : new ArrayList<>();
     }
 
+    @NonNull
     @Override
-    public RecyclerView.ViewHolder onCreateViewHolder(ViewGroup parent, int viewType) {
+    public ViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
         View itemView;
-
         // view type is to identify where to render the chat message
         // left or right
         if (viewType == SELF) {
@@ -44,36 +36,47 @@ public class ChatAdapter extends RecyclerView.Adapter<RecyclerView.ViewHolder> {
             itemView = LayoutInflater.from(parent.getContext())
                     .inflate(R.layout.chat_item_watson, parent, false);
         }
-
-
         return new ViewHolder(itemView);
     }
 
     @Override
     public int getItemViewType(int position) {
         Message message = messageArrayList.get(position);
-        if (message.getId() != null && message.getId().equals("1")) {
-            return SELF;
-        }
-
-        return position;
+        return "1".equals(message.getId()) ? SELF : WATSON;
     }
 
     @Override
-    public void onBindViewHolder(final RecyclerView.ViewHolder holder, int position) {
+    public void onBindViewHolder(@NonNull final ViewHolder holder, int position) {
         Message message = messageArrayList.get(position);
-        switch (message.type) {
-            case TEXT:
-                ((ViewHolder) holder).message.setText(message.getMessage());
-                break;
-            case IMAGE:
-                ((ViewHolder) holder).message.setVisibility(View.GONE);
-                ImageView iv = ((ViewHolder) holder).image;
-                Glide
-                        .with(iv.getContext())
-                        .load(message.getUrl())
-                        .into(iv);
-                break;
+        
+        // Always ensure message TextView is available
+        if (holder.message != null) {
+            holder.message.setVisibility(View.VISIBLE);
+            holder.message.setText(message.getMessage());
+        }
+
+        // Handle image if ImageView is available
+        if (holder.image != null) {
+            if (message.getType() == Message.Type.IMAGE) {
+                holder.message.setVisibility(View.GONE);
+                holder.image.setVisibility(View.VISIBLE);
+                String imageUrl = message.getUrl();
+                if (imageUrl != null && !imageUrl.isEmpty()) {
+                    RequestOptions options = new RequestOptions()
+                            .diskCacheStrategy(DiskCacheStrategy.ALL)
+                            .placeholder(R.drawable.placeholder_image)
+                            .error(R.drawable.error_image);
+                    
+                    Glide.with(holder.image.getContext())
+                            .load(imageUrl)
+                            .apply(options)
+                            .into(holder.image);
+                } else {
+                    holder.image.setImageResource(R.drawable.error_image);
+                }
+            } else {
+                holder.image.setVisibility(View.GONE);
+            }
         }
     }
 
@@ -82,22 +85,14 @@ public class ChatAdapter extends RecyclerView.Adapter<RecyclerView.ViewHolder> {
         return messageArrayList.size();
     }
 
-    public class ViewHolder extends RecyclerView.ViewHolder {
-        TextView message;
-        ImageView image;
+    public static class ViewHolder extends RecyclerView.ViewHolder {
+        final TextView message;
+        final ImageView image;
 
-        public ViewHolder(View view) {
+        public ViewHolder(@NonNull View view) {
             super(view);
-            message = (TextView) itemView.findViewById(R.id.message);
-            image = (ImageView) itemView.findViewById(R.id.image);
-
-            //TODO: Uncomment this if you want to use a custom Font
-            /*String customFont = "Montserrat-Regular.ttf";
-            Typeface typeface = Typeface.createFromAsset(itemView.getContext().getAssets(), customFont);
-            message.setTypeface(typeface);*/
-
+            message = view.findViewById(R.id.message);
+            image = view.findViewById(R.id.image);
         }
     }
-
-
 }
